Moved parser over to server
