+++ BLTCWeb/BLTCWeb.Client/Program.cs
@@ -5,6 +5,5 @@ using MudBlazor.Services;
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 
 builder.Services.AddMudServices();
-builder.Services.AddScoped<WebParser>();
 
 await builder.Build().RunAsync();
