+++ BLTCWeb/BLTCWeb/Components/Layout/NavMenu.razor
@@ -9,8 +9,13 @@
 <div class="nav-scrollable" onclick="document.querySelector('.navbar-toggler').click()">
     <nav class="flex-column">
         <div class="nav-item px-3">
-            <NavLink class="nav-link" href="logmanagement">
+            <NavLink class="nav-link" href="logsmanagement">
                 <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Log Management
+            </NavLink>
+        </div>
+        <div class="nav-item px-3">
+            <NavLink class="nav-link" href="dps">
+                <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Dps
             </NavLink>
         </div>
     </nav>
